I adapted width and padding for desktop, positioning of the .list-container
I modified the widths with min() to adapt it to mobile and
desktop, I just increased the padding with media query
and I moved the .list-container above the header with
z-index and translateY to make it simple.

Also I changed the background image of header with the
desktop image and adapted it with cover for mobile.